% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leanr1.R
\name{learnr1}
\alias{learnr1}
\title{Learnr tutorial 1}
\usage{
learnr1()
}
\arguments{
\item{nothing}{nothing is required}
}
\value{
Opens a RStudio or external browser containing a tutorial sheet.
}
\description{
This was taken from the leanr developers. https://github.com/rstudio/learnr/tree/master/examples
}
\examples{
learnr1()
}
