% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnr2.R
\name{learnr2}
\alias{learnr2}
\title{Learnr tutorial 2}
\usage{
learnr2()
}
\arguments{
\item{Nothing}{nothing is required}
}
\value{
Opens a RStudio or external browser containing a tutorial sheet.
}
\description{
This was taken from the leanr developers. https://github.com/rstudio/learnr/tree/master/examples
}
\examples{
learnr2()
}
